fix BUG Float label

+++ pymel/dataset/utils/helper/detach.py
@@ -102,8 +102,8 @@ def single_task_detach(
     query_y.extend([task]*len(query_dct[task]))
     
     support_x = torch.stack(support_x)
-    support_y = torch.FloatTensor(support_y)
+    support_y = torch.LongTensor(support_y)
     query_x = torch.stack(query_x)
-    query_y = torch.FloatTensor(query_y)
+    query_y = torch.LongTensor(query_y)
     
     return (support_x, support_y, query_x, query_y)
